\name{plot.predict.GORH}
\alias{plot.predict.GORH}
\title{
Plot the predicted survival curves.
}
\description{
Plot the predicted survival curve of the new subject.
}
\usage{
\method{plot}{predict.GORH}(x, ...)
}
\arguments{
  \item{x}{
An object from the predict.GORH function.
}
  \item{\dots}{
Other plot arguments.
}
}

