\name{summary.GORH}
\alias{summary.GORH}
\title{
Summary table of hypothesis tests for the coefficients in the GORH model.
}
\description{
Estimates, standard errors, test statistics and p values are presented for each coefficient in the GORH model.
}
\usage{
\method{summary}{GORH}(object, ...)
}
\arguments{
  \item{object}{
An object from the function GORH.
}
  \item{\dots}{
Other auguments to be specified.
}
}

